refactor(credential): move template in specific package
